+++ database/migrations/2023_05_29_164712_create_datos_fiscales_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('datos_fiscales', function (Blueprint $table) {
             $table->id("id_datos_fiscales");
-            $table->unsignedBigInteger("id_contrato");
+            $table->unsignedInteger("id_contrato");
             $table->string("rfc");
             $table->string("razon_social");
             $table->foreign("id_contrato")->references("id_contrato")->on("contratos")->onDelete("cascade")->onUpdate("cascade");
